Removed _queue and _ev from Recorder
The Recorder class will not be able to faithfully reproduce the queue
and ev attributes of the underlying QNode
without a significant change of the QNode due to
the removal of Observables from the queue. Because of this,
I decided to remove the corresponding properties and logic from the Recorder
class. If a discrimination is wished, it can be executed post-hoc via isinstance and
op.return_type.