Adds an operation recorder for inspecting templates (#388)
* pennylane/utils.py

* Add properties for ops and obs

* Spoof underlying QNode

Underscores were added to all attributes of the Recorder class to avoid possible
clashes with attributes of the underlying QNode. Additionally, __getattr__
was overwritten to allow the spoofing of all attributes of the
underlying QNode.

* Fix underscores and add additional attributes

The OperationRecorder class recevies new attributes:
* gates: all applied gates
* observables: all applied observables
* operations: all applied operations (gates + observables)

* Fix docstring

* Roll back additional attributes

I realized that having the recorder also record observables independently would
be complicating things too much.

* Removed _queue and _ev from Recorder

The Recorder class will not be able to faithfully reproduce the queue
and ev attributes of the underlying QNode
without a significant change of the QNode due to
the removal of Observables from the queue. Because of this,
I decided to remove the corresponding properties and logic from the Recorder
class. If a discrimination is wished, it can be executed post-hoc via isinstance and
op.return_type.

* Fix recorder.__getattr__

* Add tests

* Test context switching

* Add OperationRecorder tests

* Change __str__ implementation

* Remove unused import

* Raise error if attribute is unknown

This happens if no underlying context is specified but relaying is attempted.
We should think more about this.

* Fix test

* Add queue property and tests

* Codefactoring

* Select operations and observables post-hoc

* Update pennylane/utils.py

Co-Authored-By: Josh Izaac <josh146@gmail.com>

* Update pennylane/utils.py

Co-Authored-By: Josh Izaac <josh146@gmail.com>

* Update tests/test_utils.py

Co-Authored-By: Josh Izaac <josh146@gmail.com>

* Update tests/test_utils.py

Co-Authored-By: Josh Izaac <josh146@gmail.com>

* Update pennylane/utils.py

Co-Authored-By: Josh Izaac <josh146@gmail.com>

* Replace dict with OrderedDict

dict is only ordered starting with python 3.7

* Black

* update changelog