Add parameter batch support to the AmplitudeEmbedding template (#1761)
* Add parameter batch support to AmplitudeEmbedding

This modifies the AmplitudeEmbedding template to support a batch
of parameters for the 'features' input. The pre-processing on the
inputs is then applied for each feature_set in the batch.

* Fix redefined-argument-from-local error

* Fix wording of vector norm error message

* Add AmplitudeEmbedding to batch parameter testing

* Modify AmplitudeEmbedding test for 2d features tensor

Changes `test_throws_exception_if_features_wrong_shape` to only throw
an error if the # of dimensions of the features tensor is >2, instead
of the previous >1. This allows batched parameters to be passed.

* Remove inapplicable AmplitudeEmbedding test

The `test_throws_exception_if_features_wrong_shape` is no  longer
applicable with batching, since it verifies that the template does
not accept feature tensors with more than one dimension.

* Simplify batch check

Co-authored-by: Josh Izaac <josh146@gmail.com>

* Revert "Remove inapplicable AmplitudeEmbedding test"

This reverts commit 3fdb175cb39c854f53d8d20bba05bd64648f0239.

* Add PR & contributor to changelog

* Update doc/releases/changelog-dev.md

Co-authored-by: Josh Izaac <josh146@gmail.com>