Eliminate `requires_grad=True` warnings
Some tests supplied inputs to optimization examples that make use of
auto-differentiation. With PennyLane v0.20, function parameters not
explicitly marked with `requires_grad=True` will not be considered for
differentiation. Parameters without this attribute currently generate a
warning in anticipation of the change. Thus, all such arguments are
wrapped in a PennyLane-Numpy array with the above attribute set.
As a consequence, lists, tuples, and other data types cannot be supplied
directly as inputs in this setting.