batching ability for non-trainable inputs only following issue #2037
This function creates multiple circuit executions for batched input examples
and executes all batch inputs with the same trainable variables. The main
difference between the proposed version in the issue and this commit is the
input `argnum` this indicates the location of the given input hence gives the
ability to work across platforms.