Fix Hessian transform for JAX
Amend test to use backprop, as derivatives of order  > 1 are unsupported
by the JAX interface with parameter-shift. Additionally, change the
QNode return value to expval, as probabilities are unsupported by the
JAX interface.
Stack the tape results before determining their shape. The numpy shape
function converts its argument to ndarray first if it is not already a
single array (in this case, the tape results are a list of arrays). With
JAX, this conversion fails as the list of traced arrays cannot be
converted to a regular ndarray.