Add lightning workaround
As the lightning.qubit device inherits all properties from the
default.qubit device, the snapshot operation is currently recognized as
"supported" and the compute_matrix method called by the device.
Define `compute_matrix` for `qml.snapshot` and separately check for
the lightning device in `qml.snapshots` to raise the unsupported device
error.