Move tape result squeezing to each transform
The previous placement of the squezing inside the batch transform
assumes that all uses of the batch transform are similar to that of the
gradient transforms, which is not the case. To avoid breaking other type
of transforms with this change, the squeezing is now placed inside each
gradient transform.

Tapes constructed from QNodes now also carry the `_qfunc_output`
attribute.