Fix docker build failures
- Since the docker build has started to install the latest development
wheel of pennylane-lightning when building pennylane from master, its
dev version number cannot be reconciled by pip with the pennylane
requirement >=0.23. Changing this requirement to >=0.23.0.dev0 allows
pip to consider pre-release versions of the pennylane-lightning package
in its dependency resolution.

- Additionally, the hard requirement for semantic-version==2.6 is likely
outdated and causes another dependency issue if the package is already
installed with a higher version.