Ignore openfermionpyscf if it cannot be installed
Currently, pyscf does not provide a release for python3.10 which ships
with Ubuntu jammy (22.04LTS).
We can attempt installation of the openfermionpyscf package, but ignore
it if its installation fails (the package is only used for the
non-differentiable backend of the qml.qchem module). That way, the
package will automatically be included again in the docker build once
pyscf becomes available for python3.10 (scheduled for version 2.1).