render docs for devices.qubit functions (#4555)
I am open to notable variation in the docs here, just lmk! I mostly just
wanted to get this PR going.

**Context:**
The `devices.qubit` functions are very useful, and people should be able
to see them! They're still a little unstable, so I'm inclined to leave
those warnings in the docs for now, but they should be visible.

**Description of the Change:**
Add the `devices.qubit` qubit-simulation functions at the bottom of the
`qml.devices` sub-page in the "Internal" section (check it out
[here](https://xanaduai-pennylane--4555.com.readthedocs.build/en/4555/code/qml_devices.html#module-pennylane.devices.qubit)!)

**Benefits:**
Nice pretty docs to show users all the cool features and helper
functions we have

**Possible Drawbacks:**
It's still kinda unstable while we ramp up DQ2, but it should be ready
by the time we release 0.33