Validate wires on circuit before executing on new device; add repr to new device (#4562)
**Context:**
There are some little things that would be nice to have on the new
device. Here are 2 of them.

**Description of the Change:**
- If DQ2 has wires set on it, make sure all circuits running on the
device only use those wires
- Added a device repr similar to the old device. The only difference is
that if wires or shots are None, I exclude them from the repr. And if
they are _both_ None, then I hide the whole parenthesis component of
details.

**Benefits:**
- The new device can't mistakenly run circuits that it shouldn't be able
to
- Users see a more meaningful repr when printing a device

---------

Co-authored-by: Christina Lee <christina@xanadu.ai>