add default.qubit.legacy and rename some test devices to prepare for DQ2 (#4568)
**Context:**
DefaultQubit2 will replace DefaultQubit when users call
`qml.device("default.qubit")` soon. DefaultQubit will still be available
with the name `"default.qubit.legacy"`. My goal in this PR is to ease
that transition, because a lot of tests need changing. There are 2
reasons a test would be changed to use `default.qubit.legacy`.
1. It tests the old device
2. It fails with the new device (for better or for worse)

This PR should _only_ change tests of the first kind.

**Description of the Change:**
- Add `default.qubit.legacy` as an alternate entrypoint for
`qml.devices.DefaultQubit` (which I may now sometimes call DQL). Note
that `default.qubit` also points to DQL (for now)
- Change tests that need the old device to use `default.qubit.legacy` as
the name to help with the transition in the future.

**Benefits:**
Future PRs will be easier to review when completing the migration.

**Possible Drawbacks:**
N/A

**Related GitHub Issues:**
#4436, #4534