StateMP accepts wires (#4570)
**Context:**
In the new device interface, a device need not have wires. In the case
when it does provide wires, it would be best if measurement processes
that don't require wires at all (`state()`, `sample()`, `probs()`) know
to expand over the extra wires. To be able to do this, they need to save
the wires somewhere, and `self.wires` kinda makes sense for this:)

**Description of the Change:**
- Allow initialization of a `StateMP` with wires (now that it doesn't
indicate a density-matrix measurement).
- When given a strict subset of wires, `process_state` pads the state
with zeros for all other wire axes
- When given a different `wire_order` from `self.wires`, `process_state`
will transpose the state to match the MP wires

**Note on wire ordering:**
I use the convention that `StateMP.wires` is the global wire order we'd
like to return when used (so we can swap out measurement processes
[here](https://github.com/PennyLaneAI/pennylane/blob/078c4503b87e988f0ca68344ff57687d36f3e7fa/pennylane/devices/experimental/default_qubit_2.py#L217)),
and the `wire_order` given to `process_state` is the wire order of the
`state` being passed in (which happens
[here](https://github.com/PennyLaneAI/pennylane/blob/078c4503b87e988f0ca68344ff57687d36f3e7fa/pennylane/devices/qubit/simulate.py#L91),
at a point where we do not know about the device or its wires).

**Changes for `pad` dispatching:**
1. I'm not sure what `qml.math.pad(x, like="torch")` was even calling,
but the docs suggest to use `torch.nn.functional.pad` so I registered it
2. The tensorflow `pad` assumes padding over more than one dimension, so
I "hacked" it to add an empty broadcast dimension if needed (in other
words, turn `[pad]` into `[[pad]]`)

**Benefits:**
- `StateMP.process_state` is more powerful!
- Allows me to open a follow-up PR where DQ2 expands results for the
above listed MPs

**Possible Drawbacks:**
- Doesn't support batching (yet, I can add in this PR if necessary, just
wanted something that worked)

---------

Co-authored-by: Romain Moyard <rmoyard@gmail.com>