Update `DecompositionUndefinedError` for `Exp` class (#4571)
**Context:**
When calling the decomposition for the `Exp` class with `num_steps`
specified, an error was raised suggesting to set `num_steps`. The real
source of the error was caused because the underlying operator was not
unitary. The error message was misleading.

**Description of the Change:**
Update the decomposition function to perform added validation before
raising error.

**Benefits:**
More informative error messages.

**Possible Drawbacks:**
None

**Related GitHub Issues:**
Fixes https://github.com/PennyLaneAI/pennylane/issues/4537