Pass interface to simulate; use interface in computation (#4582)
**Context:**
We [currently cast QNode
results](https://github.com/PennyLaneAI/pennylane/blob/7c42458fd06bca3e202957733e612e33ba3a906e/pennylane/qnode.py#L978-L979)
to its interface if there are no parameters given, but this misses the
fact that a user can give builtins/vanilla-numpy parameters. In those
cases, if an interface is specified, the user would still expect
interface-specific results.

**Description of the Change:**
Ensure that QNode results are in the interface specified even when the
arguments are not. Details on how this is done:
- Added the `interface` kwarg to `simulate()`, which passes it to
`create_initial_state()`
- `create_initial_state` now casts the initial state to the interface,
even if the `StatePrepBase` operation data is not in that interface

**Benefits:**
QNode results are in the expected interface!

**Possible Drawbacks:**
final-state computation occurs in the specified interface, which may be
less performant if there is no data in that interface... but it feels
like the right thing to do. After all, it's what the user asked for