try to unbound numpy and autograd version limit (#4563)
**Context:**
We want to stop pinning numpy and just support the latest version

**Description of the Change:**

- Tweaks here and there to remove a few places where lists of arrays of
different sizes were being cast to numpy arrays. Some just in tests,
some in source code.
- Specifying `kind="stable"` for `np.argsort` to continue using the old
behaviour until we check that its alright to change it when David gets
back. Made a shortcut story to make sure this happens.
- Some signs swapped on the results of the `dipole_integral` function in
qchem, but Soran feels this is fine since all the functions that rely on
this one are still working as expected; seems to be a different but
equally valid solution. Maybe they are using `np.argsort` somewhere.
Either way, we updated the test to match the new behaviour.

**Benefits:**
Everything in our tests pass with the latest release of numpy, so we can
remove the upper bound for compatibility with PennyLane

**Possible Drawbacks:**
It appears that TensorFlow, at least on Mac, still has an upper limit
for numpy, but I guess that isn't our problem. PL works with the new
numpy. If you are using PL and tensorflow, you will have to downgrade
numpy for tensorflow.

---------

Co-authored-by: lillian542 <Lillian.frederiksen@xanadu.ai>
Co-authored-by: lillian542 <38584660+lillian542@users.noreply.github.com>