AmplitudeEmbedding inherits from StatePrep (#4583)
**Context:**
While porting DQ2, the merge_amplitude_embedding tests were failing
because `AmplitudeEmbedding` was being decomposed. I figured that since
it behaves like a `StatePrep` op, we should make it inherit from it and
gain the benefit!

**Description of the Change:**
`AmplitudeEmbedding` inherits from `StatePrep`, gaining its
`state_vector` implementation and various properties.

The only real change now that it inherits from `StatePrep` is that it
will return True for `isinstance(op, StatePrep)` checks but I think this
is not a problem.

**Benefits:**
`AmplitudeEmbedding` will behave the same on DQ2 as it did on DQL (it
would decompose once to `StatePrep`, then be "supported" by DQL, and
then it would have the custom apply function - on DQ2, it was
decomposing all the way, and then I hit a bug where it thought it could
support batching but the full decomposition disagreed). I confirmed
locally that all changed tests, `test_amplitude.py`, and
`test_merge_amplitude_embeddings.py` all passed with DQ2 as well.

**Possible Drawbacks:**
If someone somehow calls `expand_tape_state_prep` with
`AmplitudeEmbedding` and `skip_first=False`, it will decompose once to a
`StatePrep` and call it a day. This is reflected in tests, but I figured
it shouldn't be too much of a concern... users don't even have a way to
set `skip_first` to False (...yet?)