Various changes for DQ2 to work (#4534)
**Context:**
In #4436, I'm working on a mega-fix PR with far too many things to be
reviewed at once. That said, CI is now passing there with most tests
using the new device. I've copied over all changes to `pennylane/` to
this fresh PR, and if CI passes here then we should be well on our way
to swap out the device.

**Description of the Change:**
A lot of little things, but I'll try to sum them up here:
- if a device doesn't provide a `pennylane_requires`, assume it is
always supported
- use observable data (as well as the state) to determine if we should
use "backprop mode". If you have a circuit with non-trainable ops, but
the observable has trainable data, the state will still be a
non-trainable dtype at that point, and we might wrongly do CSR
dot-product instead of sum of terms.
- don't support trainable Pow ops because they don't work in backprop
(same as DQL)
- Squeeze all `CountsMP` results, not just dict results, to support
batched CountsMP results
- Use `device.preprocess` instead of changing the expansion strategy for
DQ2 in the `draw` module
- When sampling, only squeeze the last dimension with Pauli words to
properly support `1xN` ShotCopies
- don't validate jitted DiagonalQubitUnitary
- `metric_tensor` has better interface handling so it can be multiplied
by vanilla numpy arrays without breaking (needed for DQ2)
- QCut supports the new device interface

**Benefits:**
This PR is less than 200 lines and contains a significant portion of the
little things needed to make DQ2 work as the new default.qubit

**Possible Drawbacks:**
Lots of unrelated changes, might be tough to review.