`process_state` assumes flat, `state()` is always complex (#4602)
**Context:**
While porting to DQ2, we're discovering that some features are not at
parity with DQL. These are a few more. `ProbabilityMP.process_state()`
thought the input was not flat! What!!

**Description of the Change:**
- Change `ProbabilityMP.process_state` to assume the input is flat (I
just copy-pasted the contents of `marginal_prob` and used the new
`get_batch_size` helper, I didn't change `_count_samples`, Git just
thinks I did).
- Fix `StateMP.process_state` so it returns the correct shape when
batched
- Fix `StateMP.process_state` so it always returns complex values, even
if the inputs are not complex
- Update the docstring for `process_state` to state (haha) this
assumption

**Benefits:**
- Feature parity with DQL
- Better standardization, cleaner code, better code reuse