Register ParameterizedEvolution (#4598)
**Context:**
On the old device API, a `ParametrizedEvolution` uses the default
`apply_operation` behaviour if it is preferable to evolve the matrix and
apply (small matrix on a large state), but has special handling for if
the matrix gets bigger (specifically, if the wires for the matrix exceed
half the wires for the state, it evolves the state instead). Adding this
made things much faster and people were very happy.

**Description of the Change:**
Adds the same behaviour here, now based on the size of the state instead
of the number of device wires.

**Benefits:**
Presumably confers the speedups observed on DQL to DQ2.

**Possible Drawbacks:**
I tested whether the correct function was called based on the number of
calls to `math.einsum`, because `mocker.spy` can't access the
`devices.qubit.apply_operation` module due to ambiguity between the
module and the function of the same name. It's the best thing I could
come up with. It does not seem very reliable if anything changes in the
future.

---------

Co-authored-by: Matthew Silverman <matthews@xanadu.ai>
Co-authored-by: Mudit Pandey <mudit.pandey@xanadu.ai>