improve builtins shape and ndim; add `math.get_deep_interface` (#4603)
**Context:**
Sometimes people do things on lists like `[jnp.array(1.1),
jnp.array(2.2)]` and autoray thinks the interface is `builtins`. This
happens a lot with `Hamiltonian.coeffs`. I'm here to help autoray.

**Description of the Change:**
- Add `qml.math.get_deep_interface` to extract out the interface from
one of those (arbitrarily) deeply nested lists of lists, or tuples of
tuples. Note that I assume the struct isn't ragged, because if you're
asking for the interface, you're probably about to cast the whole list
and that can fail if it needs to
- Change the `builtins` implementation for `ndim` and `shape` to check
if it's a nested type containing a different interface. Then, cast to
that interface and use _its_ implementation for `ndim`/`shape`

**Benefits:**
We can get the `ndim`/`shape` of ugly nested things, and the helper
function might be generally helpful elsewhere

**Possible Drawbacks:**
None!!!

**Related GitHub Issues:**
Fixes #4601