Correct expected results on failing finite diff test (#4627)
**Context:**
A test in `finite_diff_shot_vec` was failing for the new `DefaultQubit`
for some parameters, even though the `atol` being used in comparing to
the expected results was already 0.3 (!!) for results on the order of
0.5

It turns out the analytic results were incorrect; there was an extra
factor of 2 in the derivative for probabilities, and we were comparing
the _derivative_ of the expectation value to the analytic result for the
expectation value (no derivative).

`DefaultQubitLegacy` was incorrectly passing this test. 

The problem with `DefaultQubit` was that it was _too close to the right
answer_, and just a little too far from the wrong answer being used for
comparison.

**Description of the Change:**
The analytic results have been corrected. For expectation value
derivatives, the difference between analytic and simulated values is
consistently 0.05-0.12.

Also double-checked the analytic results for the other tests in the
file; seems all good.

**Benefits:**
`DefaultQubit` (correctly) passes the test.

If something starts going wrong, we are more likely to notice, because
we set lower and more specific values for `atol` in `np.allclose`
instead of using 0.3 everywhere.

**Possible Drawbacks:**
Not a drawback per se, but there is a discrepancy between DefaultQubit
and DefaultQubitLegacy. DefaultQubitLegacy does not pass the correct
version of this test.