Update qnode_tests to use new device API (#4648)
**Context:**

All the old tests are still here: `qnode_test.py` -->
`qnode_test_legacy.py`

The `qnode_test.py` file is updated. Main changes are:

- removed tests that are only relevant for the old API
- added `stop_recording` various places in measurements to prevent
nested queueing with `default.qubit`
- add preprocessing the tape at the same place in the qnode as expanding
for DQL if `qnode.expansion_strategy = device`

The function `test_no_gradient_expansion` currently fails (marked xfail)
due to a difference in behaviour between DQL and DQ: on DQL an
unsupported operation with a defined gradient recipe is not expanded,
but on DQ2 it is. @albi3ro has added it where relevant in shortcut.