Defining custom decompositions for DQ2 (#4675)
**Context:**
The `custom_decomps` and `decomp_depth` arguments when calling
`qml.device("default.qubit")` did not work, and as such were not at
feature parity with DQL. The `set_decomposition` context also didn't
work with the new implementation of `default.qubit`

**Description of the Change:**
Kwargs `decomposer` and `max_expansion` are added to the
`preprocessing.decompose` transform. These can be specified to create
custom behaviour for decompositions.

A `custom_decomp_preprocessing` function is created, analogous to
`create_decomp_expand_fn` for DQL. It defines a custom `decomposer` for
`device.decompose` and modifies device `stopping_conditions` based on
the `custom_decomps` dictionary provided. It also modifies
`device.decompose.max_expansion` if a `decomp_depth` is specified.

There's some special treatment for controlled operators where we
override the `op.base.compute_decomposition` method for that instance of
the operator if there is a custom decomposition defined for the base.

The tests are all copy-pasted from the old device API tests, with slight
modifications where relevant for the new API, and a couple of additional
tests that seemed relevant.

**Benefits:**
Feature parity with DQL wrt custom decompositions. It works.

**Possible Drawbacks:**
It feels messy and not very nice (but it does the thing)

---------

Co-authored-by: Matthew Silverman <matthews@xanadu.ai>
Co-authored-by: albi3ro <chrissie.c.l@gmail.com>
Co-authored-by: Christina Lee <christina@xanadu.ai>
Co-authored-by: Amintor Dusko <87949283+AmintorDusko@users.noreply.github.com>
Co-authored-by: Tom Bromley <49409390+trbromley@users.noreply.github.com>