Add `DeviceDerivatives` jacobian product calculator (#4527)
This PR follows on from #4435 , and adds a child class that can handle
device-provided derivatives.

It has two private caches, `_results_cache` and `_jacs_cache`. This
handles storing the jacobian between the forward pass and the backward
pass. The caches will also store results and jacobians between
successive calls to the same instance.

The class can handle both device interfaces, `qml.Device` and
`qml.devices.Device`.

Name for the class is still up in the air. This class takes the full
jacobian from the device and then performs the dot product. The class in
#4637 takes the jacobian product itself from the device.

---------

Co-authored-by: David Wierichs <david.wierichs@xanadu.ai>
Co-authored-by: Mudit Pandey <mudit.pandey@xanadu.ai>