Lighten and Optimize Build CI (#4638)
**Context:**
Update the CI to run a "lightened" version of tests by default on pull
requests, and increase the number of groups for Jax and TF tests.

**Description of the Change:**
This PR introduces the following changes:

- A 'ligtened' test suite
- Instead of running the test suite across multiple Python versions
(3.9, 3.10, 3.11), only run the test suite against one version of Python
(3.9).

- Reduce max parallelisation for small test jobs, and assign the extra
compute room to Jax and TF tests.

- Introduction of "test durations" files to be used by pytest-split in
order to better distribute jobs across groups.
- These two files add an additional 3.1MB. At the time of writing these
two files are included as part of this PR, but it can be external and
sourced as part of CI.
- These files DO NOT need to be regularly updated, even if new JAX and
TF tests are added, as Pytest takes an avg time for tests that are not
in the durations file. If we could update them once in a while (maybe
once a year or two), that should suffice.

**Benefits:**
The changes introduced as part of this PR reduces the runtime of the
build CI. As well as consume less CI workers for PRs (by default).

Full Test Suite (Multiple Versions of Python):
Runtime:
Before: ~45min
Now:    ~25min

Workers:
Before: 30
Now:    29

Lightened Test Suite (One version of Python):
Runtime:
Before: N/A
Now:    ~15min

Workers:
Before: N/A
Now:    17

Overall, the CI on a PR that is doing a lightened test run will now
consume 43% less workers, and run 67% quicker.


**Possible Drawbacks:**
Not running multiple versions of Python on PRs does open the door for
regressions. However, if the label `ci:run-full-test-suite` is attached
to any PR, then the entire test suite runs on it. All merges to master
also run the full test suite.

**Related GitHub Issues:**
None.

---------

Co-authored-by: Qottmann <Korbinian.Kottmann@gmail.com>
Co-authored-by: Christina Lee <christina@xanadu.ai>