Decompositions for multi-controlled single qubit special unitary ops (#4697)
[sc-43590] Fixes #4473 Fixes #4660 

We previously turned off automatic decompositions for multicontrolled
single qubit gates as our decompositions only applied to special unitary
operations.

In this PR, I just check if the matrix is special unitary using its
determinant. Since this is only for single qubit gates, I'm not
concerned about the performance of this.

To decompose non-special unitary gates like `PauliZ`, we would need to
include a controlled global phase operation. With the addition of the
`GlobalPhase`, we can do so in a follow-up PR. But this one solves the
immediate problems.