Allow postselecting on mid-circuit measurements (#4604)
**Context:**
This PR adds support for postselecting on mid-circuit measurements with
`default.qubit`. The various changes to cover the edge cases of
postselection are described below.

**Description of the Change:**

Changes to measurements, math, and transforms:

* Add `postselect` kwarg to `qml.measure`.
* Add logic to `defer_measurements` to support postselection:
* Apply `qml.Projector` as an operation when postselection is requested.
* Call `broadcast_expand` when postselection is requested and batch_size
is not `None`. This is done because there is a possibility of ending up
with `NaN` values in the state, and we only want the state vector of the
"bad" parameter to have `NaN` values, not the entire state vector.
* Raise an error if the device used for executing the deferred tape is
not `default.qubit`.
* Added tensorflow single dispatch to `qml.math` for `isnan`, `full`,
and `any`. Why must tensorflow name their functions slightly differently
than everyone else? The functions I needed to add dispatches for are
called `tf.math.is_nan`, `tf.fill`, and `tf.reduce_any` respectively 😢

Changes to `default.qubit`:

* Added logic for applying `qml.Projector` and renormalizing state in
`simulate.py`:
* If the probability of the postselected state being measured is zero or
extremely close to zero, we make it exactly zero and let execution
continue so that the results become invalid rather than incorrect.
* Use `np.random.binomial` to update the number of shots on the tape
such that only a subset of the shots are then used. This is done to
simulate behaviour on real hardware, where invalid samples would be
discarded.
* Added `_FlexShots` class which inherits `Shots` to `simulate.py` that
allows shot vectors to contain 0 shots. We need this in cases when the
probability of measuring the postselected state is close to 0 but not 0.
In these cases, it is possible to end up with 0 samples that are valid,
and we need to be able to handle that.
* Added logic to `sampling.py` for catching errors raised by the rng if
attempting to generate samples from nan probabilities. When this
happens, samples of shape (`shots, len(wires))` with NaN values are
returned.

The following is done to measurements in cases where the state is
invalid or the number of shots is 0:
* Measurement processes return `NaN` values within their expected shape.
Eg, `qml.probs` returns a `2**n` size array with all `NaN`s, and
`qml.expval` returns a single `NaN`.

* QInfo measurements don't work for analytic execution, and `probs` and
`counts` don't work for finite shots if state has `NaN` values.

Other changes:
* Removed application of `defer_measurements` in QNode construction if
the device is `default.qubit`. This is done so that multiple tapes are
not created by the transform in case there is postselection and
broadcasting.
* Added error to drawer for circuits with mid-circuit measurements.

TODOs:
* [x] Update documentation and examples to include info about possible
invalid results. This should include info about:
* [x] when invalid results can occur: when the postselection probability
is zero, or when the postselection probability is so low that updating
the number of shots per the binomial distribution leads to zero shots.
  * [x] Which measurements are invalid,
* [x] Add integration tests.
* [x] Fix sampling tests. In an integrated setting, postselection works
with finite shots measurements except for probs and counts, but the unit
tests just fail. I haven't figured out yet what needs to change. All
tests related to postselection are in
`tests.devices.qubit.test_sampling.TestInvalidStateSamples`.

**Benefits:**
Postselection is available on `default.qubit`.

**Possible Drawbacks:**
* Postselection with some measurements causes errors that aren't
necessarily easy to understand.
* I had to remove the use of `defer_measurements` from qnode
construction if the new device was being used as postselection would
split tapes if the tape was broadcasted. This causes the drawer to not
work with `default.qubit` for circuits with mid-circuit measurements
unless `defer_measurements` is explicitly applied to the QNode.

**Related GitHub Issues:**

[Here's a
gist](https://gist.github.com/mudit2812/f7c83190ba69dcf03af5f589a2bad6a8)
with the notebook I've been using to run examples locally. It has flags
to test analytic vs finite shots, integer shots vs shot vector, exactly
zero vs close to zero postselected state probability, and unit vs
integration execution

---------

Co-authored-by: Matthew Silverman <matthews@xanadu.ai>
Co-authored-by: Romain Moyard <rmoyard@gmail.com>
Co-authored-by: Edward Jiang <34989448+eddddddy@users.noreply.github.com>
Co-authored-by: Christina Lee <christina@xanadu.ai>
Co-authored-by: Edward Jiang <eddy3.14159@gmail.com>
Co-authored-by: soranjh <40344468+soranjh@users.noreply.github.com>
Co-authored-by: Jay Soni <jbsoni@uwaterloo.ca>
Co-authored-by: Vincent Michaud-Rioux <vincentm@nanoacademic.com>
Co-authored-by: Utkarsh <utkarshazad98@gmail.com>
Co-authored-by: Diego <67476785+DSGuala@users.noreply.github.com>
Co-authored-by: lillian542 <Lillian.frederiksen@xanadu.ai>
Co-authored-by: lillian542 <38584660+lillian542@users.noreply.github.com>
Co-authored-by: Rashid N H M <95639609+rashidnhm@users.noreply.github.com>
Co-authored-by: Stepan Fomichev <stepan.fomichev95@gmail.com>
Co-authored-by: Tom Bromley <49409390+trbromley@users.noreply.github.com>