Always draw with device wires first (#4729)
**Context:**
Sometimes, tapes will use standard wires (eg. `[0, 1, 2]`) but will
report them out of order, depending on the order in which they are used.
If we are drawing a qnode (in other words, a circuit bound to a device),
we have the true wire order available to us.

**Description of the Change:**
Use the device wire order wherever possible in drawing logic

**Benefits:**
We don't accidentally draw in funky wire orders like `[1, 0]`. See the
docstring that I changed in `tape_expand` as an example (before this PR,
it would swap wires without the `wire_order` specified)

**Possible Drawbacks:**
N/A

**Related GitHub Issues:**
first found in #4711