Update counts to handle floats/NaN values from hardware (#4739)
The QuEra hardware sometimes has failed measurements, which return as
NaN. Because of this, all samples from the device are returned with type
`float`, and some of them are `np.NaN`. These were not handled well in
counts. On the `_default_qubit` device, nothing was converted to
integers, so it did this:

```
{'0.00.00.0': 149,
'0.00.01.0': 281, 
'0.00.0nan': 2,
'0.01.00.0': 277,
'0.01.01.0': 4,
'0.0nan0.0': 4,
'1.00.00.0': 268,
'1.00.0nan': 2,
'1.01.00.0': 5,
'nan0.00.0': 8}
```

We don't like that. Now they are converted into integers if there is no
observable, just like in `CountsMP`.

Also added a line both there and in `CountsMP` to avoid converting the
`NaN` values to 0 and implying they were ground state measurements -
instead, any sample where at least one wire failed is disregarded when
totalling the number of counts.

Has been tested with the QuEra hardware, and now returns nicely:

```
>>> res = circuit(params)
>>> res
{'000': 230, '001': 251, '010': 252, '011': 1, '100': 249, '110': 4}
```

with 13 disregarded failed measurements:
```
>>> sum(count for count in res.values())
987
```