Pinned jax and tensorflow versions for external-libraries-tests (#4717)
**Context:**
Fixes few CI issues that was causing build failures.

- Pinned versions of jax/tensorflow being installed for
`external-libraries-tests`
- Removed keras_version input

**Description of the Change:**
As stated in the context, this PR introduces to changes.

`Pinned versions of jax/tensorflow being installed for
`external-libraries-tests`
Previously, this was not pinned, causing CI to always installed the
latest available version of these packages from Pypi. Which was causing
sporradic test failures for the job.

`Removed keras_version input`
THis was an input into `unit-test.yml` .. but it was never used as we
always install the same version of tensorflow/keras. Therefore, it'd
cause not confusion in the future in having that input.

**Benefits:**
- More deterministic CI builds.

**Possible Drawbacks:**
None.

**Related GitHub Issues:**
None. [Shortcut
Story](https://app.shortcut.com/xanaduai/story/48505/fix-incorrect-versions-of-jax-tensorflow-being-installed-for-external-libraries-tests-on-pennylane-ci)

Co-authored-by: Christina Lee <christina@xanadu.ai>