deprecate tape.is_sampled and tape.all_sampled (#4773)
**Context:**
`tape.is_sampled` and `tape.all_sampled` are not very useful and could
really be evaluated on a per-call basis. Plus, maintaining the
correctness of these values can be difficult if new measurements are
introduced.

**Description of the Change:**
Raise a warning whenever those properties are accessed.

**Benefits:**
We're getting rid of stuff that may not necessarily belong on the tape
API! 🧹

**Possible Drawbacks:**
In the deprecation notice, we're suggesting the 4 measurement types that
are currently being used, but they are theoretically subject to change.
We discussed other things to recommend, but I think this is the best
option so far because it's the same as what we use in
`QubitDevice.execute`. Sure, it might change one day. But by then, the
properties (and their warnings) will be removed altogether!

[sc-45870]