Deprecate Observable.return_type (#4762)
**Context:**
`Observable.return_type` is a relic of an older PennyLane, and should be
removed.

**Description of the Change:**
- raise a warning whenever `Observable.return_type` is accessed
- Observable reprs (like Tensor) will never try to add the return type
to it
- change CircuitGraph to expect measurements instead of observables
- funky stuff with `tape._par_info[x]` -> I kept the info["op"] to be
the observable (and not the surrounding measurement) because a lot of
code expects it to be an operator (fair enough!). whenever code needed
the return type, I instead get the `op_idx` from the info, then access
the tape at that index to get the measurement
- `var_param_shift` now checks if it's replacing a trainable hamiltonian
and fails before passing to `expval_param_shift`. This was a weird one:
`expval_param_shift` would receive something like `qml.expval(H)`, but
`H.return_type` would still be `Variance` because we never updated it
when swapping measurements. The new way is to just fail right away.

**Benefits:**
- the property has been deprecated!
- Kept it as minimal as possible while avoiding using the property in PL
source

**Possible Drawbacks:**
To keep things minimal, there are still some arguments named
"observables" when they are actually supposed to be measurements. I can
change that too if we want.