add a setter to avoid breaking Observable.return_type (#4798)
**Context:**
Some plugin tests failed because they were setting this property, which
I did not expect.

**Description of the Change:**
Add a setter to the `Observable.return_type` property to avoid breaking
things in this release. Also warn if someone uses it.

**Benefits:**
We don't break `Observable.return_type` in this release.

**Possible Drawbacks:**
People can still set it? idk seems like a safe call for now. I'm not
validating the type provided

**Related GitHub Issues:**
missed in #4762